Upgrade to Kotlin 2
This commit upgrade to Kotlin 2.1. Two related dependencies have been
updated as well: Kotlin Coroutines to 1.9.0, and Kotlin Serialization to
1.7.3.

As of Kotlin 2, it is now longer possible to have a Java type and a
Kotlin type with the same name. As our code samples follow that
unfortunate pattern, this commit makes sure that the Kotlin sample code
does not depend on any of the Java counterpart and configure the kotlin
compilation plugin to ignore Java sources.

The minimum version of Gradle is 7.6.4. It bundles a version of Kotlin that
cannot compile a Kotlin build script when spring-core, compiled with Kotlin
2.1, is on the classpath. Using Gradle 8.12 to run the DSL tests avoids
the problem.

Co-authored-by: Andy Wilkinson <andy.wilkinson@broadcom.com>